mi_api_rest/requirements.txt
DJango==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.5
django-cors-headers==3.14.0
drf-yasg==1.21.7